Remove vim-script plugin
When I execute 'PluginInstall' in Seoul DeskTop,
The install was stopped on that plugin.

Also Vundle doesn't need it anymore.

In https://vim-scraper.github.io/
"Now that most Vim scripts are developed on GitHub and installable straight from the source,
I'm happy to say that there's no need for vim-scripts anymore"

Signed-off-by: Sunwook Eom <sunwook5492@gmail.com>